change: Rename Topology module accordingly to their Components
Rename Topology to HostsTopologyChart and TopologyDisplay to
HostsTopologyDisplay accordingly to their components inside the modules.
